feat(view): standalone outbound file drag on Windows + SDL native-handle plumbing
A standalone Pulp app (tree owned by a WindowHost, not a plugin view host)
could not drag files out: View::start_file_drag falls through to the free
begin_file_drag(native_view, request), which was a no-op everywhere except
macOS. Two gaps closed for Windows:

1. The SDL WindowHost exposed no native handle, so start_file_drag bailed at
   the null-handle guard before reaching any backend. Add
   SdlWindowHostImpl::native_content_view_handle() returning the HWND (Windows)
   / X11 Window id (Linux) via SDL3 window properties.
2. No free begin_file_drag backend off macOS. Extract the OLE drag source
   (CF_HDROP IDataObject + IDropSource + DoDragDrop) from plugin_view_host_win.cpp
   into the Skia-free shared header win_file_drag.hpp, and add drag_drop_win.cpp
   providing begin_file_drag() -- compiled unconditionally on _WIN32 (OLE needs
   no GPU), so the drag_drop.cpp no-op now guards "not APPLE and not WIN32". The
   plugin host delegates to the same win_drag::win_run_file_drag(), so both
   paths share one backend.

Linux standalone outbound (the SDL X11 handle is already plumbed) follows next:
its begin_file_drag needs the XDND source extracted from run_xdnd_source, so
it stays the drag_drop.cpp no-op here (returns false, no regression).
